perf(fireworks): executing fadeout to the whole firework instead of each particle
