+++ src/components/FireworkParticle.tsx
@@ -20,14 +20,7 @@ export interface FireworkParticleProps {
 }
 
 export const FireworkParticle = memo(
-  ({
-    x,
-    y = 0,
-    finalPos,
-    r = 8,
-    color,
-    autoHide = true,
-  }: FireworkParticleProps) => {
+  ({ x, y = 0, finalPos, r = 8, color }: FireworkParticleProps) => {
     const cx = useTiming(
       { to: finalPos.x, from: x },
       { duration: 1200, easing: Easing.linear }
@@ -38,21 +31,11 @@ export const FireworkParticle = memo(
       { duration: 1200, easing: Easing.linear }
     );
 
-    const opacityValue = useTiming(
-      { to: 0, from: 1 },
-      { duration: 2000, easing: Easing.linear }
-    );
-
     const transformCircle = useComputedValue(
       () => vec(cx.current, cy.current),
       [cx, cy]
     );
 
-    const opacity = useComputedValue(
-      () => (autoHide ? opacityValue.current : 1),
-      [opacityValue]
-    );
-
-    return <Circle c={transformCircle} r={r} color={color} opacity={opacity} />;
+    return <Circle c={transformCircle} r={r} color={color} />;
   }
 );
